added test_utils for simplifying tests
