+++ core/tests/Test_ContainerTraits.hpp
@@ -90,77 +90,6 @@ void test_container_traits() {
                        typename std::add_const<Container>::type>::type>::value);
 }
 
-template <typename Container>
-void test_traits(Morpheus::DenseVectorTag) {
-  static_assert(std::is_same<typename Container::value_array_pointer,
-                             typename std::add_pointer<
-                                 typename Container::value_type>::type>::value);
-  static_assert(std::is_same<typename Container::value_array_reference,
-                             typename std::add_lvalue_reference<
-                                 typename Container::value_type>::type>::value);
-}
-
-template <typename Container>
-void test_traits(Morpheus::DenseMatrixTag) {
-  static_assert(std::is_same<typename Container::value_array_pointer,
-                             typename std::add_pointer<
-                                 typename Container::value_type>::type>::value);
-  static_assert(std::is_same<typename Container::value_array_reference,
-                             typename std::add_lvalue_reference<
-                                 typename Container::value_type>::type>::value);
-}
-
-template <typename Container>
-void test_traits(Morpheus::CooTag) {
-  static_assert(std::is_same<typename Container::index_array_pointer,
-                             typename std::add_pointer<
-                                 typename Container::index_type>::type>::value);
-  static_assert(std::is_same<typename Container::index_array_reference,
-                             typename std::add_lvalue_reference<
-                                 typename Container::index_type>::type>::value);
-
-  static_assert(std::is_same<typename Container::value_array_pointer,
-                             typename std::add_pointer<
-                                 typename Container::value_type>::type>::value);
-  static_assert(std::is_same<typename Container::value_array_reference,
-                             typename std::add_lvalue_reference<
-                                 typename Container::value_type>::type>::value);
-}
-
-template <typename Container>
-void test_traits(Morpheus::CsrTag) {
-  static_assert(std::is_same<typename Container::index_array_pointer,
-                             typename std::add_pointer<
-                                 typename Container::index_type>::type>::value);
-  static_assert(std::is_same<typename Container::index_array_reference,
-                             typename std::add_lvalue_reference<
-                                 typename Container::index_type>::type>::value);
-
-  static_assert(std::is_same<typename Container::value_array_pointer,
-                             typename std::add_pointer<
-                                 typename Container::value_type>::type>::value);
-  static_assert(std::is_same<typename Container::value_array_reference,
-                             typename std::add_lvalue_reference<
-                                 typename Container::value_type>::type>::value);
-}
-
-template <typename Container>
-void test_traits(Morpheus::DiaTag) {
-  static_assert(std::is_same<typename Container::index_array_pointer,
-                             typename std::add_pointer<
-                                 typename Container::index_type>::type>::value);
-  static_assert(std::is_same<typename Container::index_array_reference,
-                             typename std::add_lvalue_reference<
-                                 typename Container::index_type>::type>::value);
-
-  static_assert(std::is_same<typename Container::value_array_pointer,
-                             typename std::add_pointer<
-                                 typename Container::value_type>::type>::value);
-  static_assert(std::is_same<typename Container::value_array_reference,
-                             typename std::add_lvalue_reference<
-                                 typename Container::value_type>::type>::value);
-}
-
 template <typename Container, typename Tag, typename ValueTypeRef,
           typename IndexTypeRef, typename LayoutRef, typename SpaceRef>
 void test_traits() {
